remove prepare statement, fix wrong db method in create and update

+++ internal/customer/repositories/customer_repository.go
@@ -9,7 +9,7 @@ import (
 type CustomerRepository interface {
 	Create(ctx context.Context, data model.TCustomers) (int64, error)
 	FindById(ctx context.Context, id int64) (*model.TCustomers, error)
-	UpdateBalance(ctx context.Context, id int64, balance float64) (float64, error)
+	UpdateBalance(ctx context.Context, id int64, balance float64) (int64, error)
 	GetBalance(ctx context.Context, id int64) (*model.CustomerBalance, error)
 	InquiryBalance(ctx context.Context, id int64) (float64, error)
 	FindByEmail(ctx context.Context, email string) (*model.TCustomers, error)
